General additions to board as well as almost functional implementation of console chess.
Corrected initialization of the board classes.

Removed comments from the generic board class.

Added a lot of functionality to the generic class, including:
	- Winner conditions,
	- Checking stalemate and checkmate,
	- Upgrading a piece,
	- Selecting a piece,
	- Forfeiting a match

Hardened the code to have lower chance of nullptr appearing inside BoardState instance of a board.
Also added more null checks, just in case.

Reduced the complexity of the logic of tryMove method.
The method is most definitely slower than before, but is much more linear and easier to reason about.

Moved the didMove variable from piece into PieceStorage.

Added new method to generic piece: getAllThreatningMoves.
This method removes all positions that the piece threatens. This is for a special case of pawn being able to move
forwards but not threatening these positions. King would've been unable to move in front of an enemy pawn otherwise.

GenericPiece::move now returns a pair of bool and PieceStorage.
The storage returned is the piece that was taken. In case a pawn captures a shadow pawn, the returned storage
is the storage for the pawn itself, instead of the shadow pawn.

Added a Pat color, representing a Pat for the winner. The playing player's color is never set to pat.

Changed threat back to an array of 2 unsigned integers, as the references to threatening pieces is actually not that useful.
Albeit it might be reverted again and used for rendering of pieces in the graphical part.

Added initial implementation of console control for chess.
The class is almost completed, will require a little more work for some more flags as well as implementation of
turns inside the generic board.