Finished the commands for console chess ui, except turn.
Moved and template-tized upgrade functions in GenericBoard, making it possible to pass stateful lambdas into them now.

Added profiling to the codebase!
Profiling can be enabled by using profile command in the console chess. This command will start taking notes of execution time
of certain functions, as well as itself, whcih will be then put into a file at the end of the program, or when profiling is
turned off.

Fixed the _canDoMove method returning moves to check in regards to the playing player instead of the piece state.